requirements.txt
# Core video processing
moviepy>=1.0.3
requests>=2.32.0
numpy>=1.26.0
Pillow>=10,<12
faster-whisper>=1.0.3

# Web interface
Flask>=2.3.0
pywebview>=4.0.0

# Authentication and database
flask-sqlalchemy>=3.1.1
flask-login>=0.6.3
flask-wtf>=1.2.2
python-dotenv>=1.1.1

# YouTube API integration (optional)
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0

# Production deployment
gunicorn>=21.2.0
psycopg2-binary>=2.9.0

